requirements.txt
Django==5.0.3
djangorestframework==3.15.1
psycopg-binary
dj-database-url==2.1.0
python-dotenv==1.0.1
groq==1.1.2
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
